show menu option to everyone

+++ index.php
@@ -40,7 +40,9 @@ if(permission_exists('linphone_manage_all')) {
         echo button::create(['type'=>'button','label'=>"show all",'icon'=>$_SESSION['theme']['button_icon_all'],'link'=>'?show=all']);
     }
 }
-echo button::create(['type'=>'button','label'=>"New",'icon'=>$_SESSION['theme']['button_icon_add'],'id'=>'btn_add','name'=>'btn_add','link'=>'device_edit.php']);
+if(permission_exists('linphone_manage_domain') || permission_exists('linphone_manage_all')) { // edit page doesn't currently support linphone_manage_self
+    echo button::create(['type'=>'button','label'=>"New",'icon'=>$_SESSION['theme']['button_icon_add'],'id'=>'btn_add','name'=>'btn_add','link'=>'device_edit.php']);
+}
 echo "	</div>\n";
 echo "	<div style='clear: both;'></div>\n";
 echo "</div>\n";
